+++ CHANGELOG.md
@@ -4,10 +4,14 @@
 
 ## Features
 
+- Added filters for custom data to the sidebar tree.
+- Files may be deleted permanently instead of being sent to the trash folder - useful to save disk space or if trashing is not supported on your system.
 - Medley sections are now checked if they start on a line start and end on a line end, otherwise a warning is issued.
+- The song previewer can be launched as a standalone window (`usdb_syncer preview path/to/song.txt`).
 
 ## Fixes
 
+- Fixed deleting songs on startup that were removed from USDB.
 - The Syncer will now also retry for Vimeo resources with logged-in cookies if necessary.
 
 <!-- 0.15.0 -->